@@ -5,7 +5,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.2.0+2
 
 environment:
-  sdk: '>=2.7.6 <3.0.0'
+  sdk: '>=2.18.6 <3.0.0'
 
 dependencies:
   flutter:
