@@ -4,11 +4,11 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:healthcare_mania_legacy_new/screens/model_list_screen.dart';
 //import 'package:flutter_stetho/flutter_stetho.dart';
 
-void main() async {runApp(const MyApp());
+void main() async {runApp( MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
 
   @override
